requirements.txt
altair==1.0.0
auto-sklearn==0.1.0
ConfigArgParse==0.11.0
ConfigSpace==0.2.1
cycler==0.10.0
Cython==0.24.1
dask==0.11.1
decorator==4.0.10
docutils==0.12
entrypoints==0.2.2
ipykernel==4.5.0
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
Jinja2==2.8
joblib==0.10.2
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.0
liac-arff==2.1.0
linecache2==1.0.0
lockfile==0.12.2
MarkupSafe==0.23
matplotlib==1.5.3
mistune==0.7.3
mock==2.0.0
nbconvert==4.2.0
nbformat==4.1.0
networkx==1.11
nose==1.3.7
notebook==4.2.3
numpy==1.11.2
pandas==0.19.0
pbr==1.10.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.4.1
prompt-toolkit==1.0.7
protobuf==3.0.0
psutil==4.3.1
ptyprocess==0.5.1
Pygments==2.1.3
pynisher==0.4.2
pyparsing==2.1.10
pyrfr==0.2.0
python-dateutil==2.5.3
pytz==2016.7
PyYAML==3.12
pyzmq==16.0.0
qtconsole==4.2.1
scikit-image==0.12.3
scikit-learn==0.18
scipy==0.18.1
simplegeneric==0.8.1
six==1.10.0
smac==0.0.1
tensorflow==0.11.0rc0
terminado==0.6
toolz==0.8.0
tornado==4.4.2
traceback2==1.4.0
traitlets==4.3.1
unittest2==1.1.0
vega==0.4.3
wcwidth==0.1.7
widgetsnbextension==1.2.6
xgboost==0.4a30
